+++ src/models/events.ts
@@ -56,6 +56,10 @@ export interface Event {
    */
   conferencing: Conferencing;
   /**
+   * Visibility of the event, if the event is private or public.
+   */
+  visibility: Visibility;
+  /**
    * Description of the event.
    */
   description?: string;
@@ -92,24 +96,20 @@ export interface Event {
   /**
    * Organizer of the event.
    */
-  organizer: EmailName;
+  organizer?: EmailName;
   /**
    * An list of RRULE and EXDATE strings.
    * @see <a href="https://datatracker.ietf.org/doc/html/rfc5545#section-3.8.5">RFC-5545</a>
    */
   recurrence?: string[];
   /**
-   * List of reminders for the event.
+   * A list of reminders to send for the event. If left empty or omitted, the event uses the provider defaults.
    */
-  reminders?: Reminder[];
+  reminders?: Reminders;
   /**
    * Status of the event.
    */
   status?: Status;
-  /**
-   * Visibility of the event, if the event is private or public.
-   */
-  visibility?: Visibility;
 }
 
 /**
@@ -147,14 +147,9 @@ export interface CreateEventRequest {
    */
   conferencing?: Conferencing;
   /**
-   * The number of minutes before the event start time when a user wants a reminder for this event.
-   * Reminder minutes need to be entered in the following format: "[20]".
+   * A list of reminders to send for the event. If left empty or omitted, the event uses the provider defaults.
    */
-  reminderMinutes?: string;
-  /**
-   * Method to remind the user about the event. (Google only).
-   */
-  reminderMethod?: string;
+  reminders?: Reminders;
   /**
    *  A list of key-value pairs storing additional data.
    */
@@ -212,7 +207,6 @@ export interface ListEventQueryParams extends ListQueryParams {
    * Different providers have different semantics for cancelled events.
    */
   showCancelled?: boolean;
-  eventId?: string;
   /**
    * Specify calendar ID of the event. "primary" is a supported value indicating the user's primary calendar.
    */
@@ -324,7 +318,7 @@ type RsvpStatus = 'yes' | 'no' | 'maybe';
 /**
  * Enum representing the visibility of an event.
  */
-type Visibility = 'public' | 'private';
+type Visibility = 'default' | 'public' | 'private';
 
 /**
  * Enum representing the supported conferencing providers.
@@ -549,7 +543,22 @@ export interface Participant {
 /**
  * Interface representing the reminders field of an event.
  */
-export interface Reminder {
+export interface Reminders {
+  /**
+   * Whether to use the default reminders for the calendar.
+   * When true, uses the default reminder settings for the calendar
+   */
+  useDefault: boolean;
+  /**
+   * A list of reminders for the event if useDefault is set to false.
+   */
+  overrides: ReminderOverride[];
+}
+
+/**
+ * Interface representing the reminder details for an event.
+ */
+export interface ReminderOverride {
   /**
    * The number of minutes before the event start time when a user wants a reminder for this event.
    * Reminder minutes are in the following format: "[20]".