+++ src/models/messages.ts
@@ -1,6 +1,6 @@
 import { EmailName } from './events.js';
-import { CreateFileRequest, File } from './files.js';
 import { ListQueryParams } from './listQueryParams.js';
+import { Attachment, CreateAttachmentRequest } from './attachments.js';
 
 /**
  * @internal Internal interface for creating a message.
@@ -25,21 +25,11 @@ export interface BaseCreateMessage {
   /**
    * An array of files to attach to the message.
    */
-  attachments?: CreateFileRequest[];
-  /**
-   * A short snippet of the message body.
-   * This is the first 100 characters of the message body, with any HTML tags removed.
-   */
-  snippet?: string;
+  attachments?: CreateAttachmentRequest[];
   /**
    * The message subject.
    */
   subject?: string;
-  /**
-   * A reference to the parent thread object.
-   * If this is a new draft, the thread will be empty.
-   */
-  threadId?: string;
   /**
    * The full HTML message body.
    * Messages with only plain-text representations are up-converted to HTML.
@@ -49,10 +39,6 @@ export interface BaseCreateMessage {
    * Whether or not the message has been starred by the user.
    */
   starred?: boolean;
-  /**
-   * Whether or not the message has been read by the user.
-   */
-  unread?: boolean;
 }
 
 /**
@@ -87,7 +73,21 @@ export interface BaseMessage extends Omit<BaseCreateMessage, 'attachments'> {
   /**
    * An array of files attached to the message.
    */
-  attachments?: File[];
+  attachments?: Attachment[];
+  /**
+   * A short snippet of the message body.
+   * This is the first 100 characters of the message body, with any HTML tags removed.
+   */
+  snippet?: string;
+  /**
+   * A reference to the parent thread object.
+   * If this is a new draft, the thread will be empty.
+   */
+  threadId?: string;
+  /**
+   * Whether or not the message has been read by the user.
+   */
+  unread?: boolean;
 }
 
 /**