+++ tests/resources/messages.spec.ts
@@ -1,7 +1,7 @@
 import APIClient from '../../src/apiClient';
 import { Messages } from '../../src/resources/messages';
-import { CreateFileRequest } from '../../src/models/files';
 import { createReadableStream, MockedFormData } from '../testUtils';
+import { CreateAttachmentRequest } from '../../src/models/attachments';
 jest.mock('../src/apiClient');
 
 // Mock the FormData constructor
@@ -153,7 +153,7 @@ describe('Messages', () => {
         subject: 'This is my test email',
       };
       const fileStream = createReadableStream('This is the text from file 1');
-      const file1: CreateFileRequest = {
+      const file1: CreateAttachmentRequest = {
         filename: 'file1.txt',
         contentType: 'text/plain',
         content: fileStream,